# v-sign-be/openapi-backend.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/quiz-attempts/{attemptId}/submit" : {
      "post" : {
        "tags" : [ "quiz-attempt-controller" ],
        "operationId" : "submit_1",
        "parameters" : [ {
          "name" : "attemptId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SubmitAttemptRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseQuizResultResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/change-password" : {
      "post" : {
        "tags" : [ "profile-controller" ],
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/request" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "requestPasswordReset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments/{assessmentId}/submissions" : {
      "post" : {
        "tags" : [ "assessment-controller" ],
        "operationId" : "submit_2",
        "parameters" : [ {
          "name" : "assessmentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssessmentSubmissionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAssessmentSubmissionResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/profile" : {
      "patch" : {
        "tags" : [ "profile-controller" ],
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payments/{transactionId}" : {
      "patch" : {
        "tags" : [ "admin-payment-controller" ],
        "operationId" : "overrideStatus",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAdminPaymentRecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quiz-attempts/{attemptId}/review" : {
      "get" : {
        "tags" : [ "quiz-attempt-controller" ],
        "operationId" : "review",
        "parameters" : [ {
          "name" : "attemptId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseQuizReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{transactionId}" : {
      "get" : {
        "tags" : [ "payment-controller" ],
        "operationId" : "status",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePaymentStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assessments/{assessmentId}" : {
      "get" : {
        "tags" : [ "assessment-controller" ],
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "assessmentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAssessmentDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users" : {
      "get" : {
        "tags" : [ "admin-user-controller" ],
        "operationId" : "listUsers",
        "parameters" : [ {
          "name" : "role",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAdminUserListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payments" : {
      "get" : {
        "tags" : [ "admin-payment-controller" ],
        "operationId" : "listPayments",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAdminPaymentPageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/kpis" : {
      "get" : {
        "tags" : [ "admin-kpi-controller" ],
        "operationId" : "kpis",
        "parameters" : [ {
          "name" : "fromDate",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "toDate",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAdminKpiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "QuizAnswerRequest" : {
        "required" : [ "questionId", "selectedAnswerId" ],
        "type" : "object",
        "properties" : {
          "questionId" : {
            "type" : "string"
          },
          "selectedAnswerId" : {
            "type" : "string"
          }
        }
      },
      "SubmitAttemptRequest" : {
        "type" : "object",
        "properties" : {
          "answers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QuizAnswerRequest"
            }
          },
          "durationSeconds" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "QuizResultResponse" : {
        "type" : "object",
        "properties" : {
          "attemptId" : {
            "type" : "string"
          },
          "score" : {
            "type" : "integer",
            "format" : "int32"
          },
          "passed" : {
            "type" : "boolean"
          },
          "xpAwarded" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reviewAvailable" : {
            "type" : "boolean"
          },
          "timedOut" : {
            "type" : "boolean"
          },
          "unansweredCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SuccessResponseQuizResultResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/QuizResultResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ChangePasswordRequest" : {
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object",
        "properties" : {
          "currentPassword" : {
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 2147483647,
            "minLength" : 8,
            "type" : "string"
          }
        }
      },
      "SuccessResponseVoid" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "RegisterRequest" : {
        "required" : [ "email", "password" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 8,
            "type" : "string"
          },
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "displayName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "AuthResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/AuthUserResponse"
          }
        }
      },
...
      "SuccessResponseAuthResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AuthResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PasswordResetRequest" : {
        "required" : [ "email" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          }
        }
      },
      "LoginRequest" : {
        "required" : [ "email", "password" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },
      "AnswerRequest" : {
        "required" : [ "questionId" ],
        "type" : "object",
        "properties" : {
          "questionId" : {
            "type" : "string"
          },
          "selectedOptionId" : {
            "type" : "string"
          },
          "selectedAnswerId" : {
            "type" : "string"
          }
        }
      },
      "AssessmentSubmissionRequest" : {
        "required" : [ "userId" ],
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "answers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AnswerRequest"
            }
          }
        }
      },
      "AssessmentSubmissionResultResponse" : {
        "type" : "object",
        "properties" : {
          "assessmentId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "score" : {
            "type" : "integer",
            "format" : "int32"
          },
          "passed" : {
            "type" : "boolean"
          },
          "correctAnswers" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalQuestions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "awardedXp" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SuccessResponseAssessmentSubmissionResultResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AssessmentSubmissionResultResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UpdateProfileRequest" : {
        "type" : "object",
        "properties" : {
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "displayName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "avatarUrl" : {
            "pattern" : "^$|^https?://.+",
            "type" : "string"
          },
          "bio" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
...
      "SuccessResponseProfileResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ProfileResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ManualPaymentStatusRequest" : {
        "required" : [ "reason", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 10,
            "type" : "string"
          }
        }
      },
      "AdminPaymentRecordResponse" : {
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string"
          },
          "userEmail" : {
            "type" : "string"
          },
          "planId" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          },
          "overrideReason" : {
            "type" : "string"
          }
        }
      },
      "SuccessResponseAdminPaymentRecordResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPaymentRecordResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseQuizReviewResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/QuizReviewResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponsePaymentStatusResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PaymentStatusResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseAssessmentDetailResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AssessmentDetailResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseAdminUserListResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminUserListResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseAdminPaymentPageResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPaymentPageResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseAdminKpiResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminKpiResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}